+++ backend/rush/http/article/article-delete.http
