+++ backend/rush/src/main/java/rush/rush/controller/comment/CommentLikeController.java
@@ -1,4 +1,4 @@
-package rush.rush.controller.article.comment;
+package rush.rush.controller.comment;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 import rush.rush.domain.MapType;
 import rush.rush.security.CurrentUser;
 import rush.rush.security.user.UserPrincipal;
-import rush.rush.service.article.comment.CommentLikeService;
+import rush.rush.service.comment.CommentLikeService;
 
 @RestController
 @RequiredArgsConstructor
