+++ backend/rush/src/main/java/rush/rush/controller/comment/DeleteCommentController.java
@@ -1,4 +1,4 @@
-package rush.rush.controller.article.comment;
+package rush.rush.controller.comment;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import rush.rush.security.CurrentUser;
 import rush.rush.security.user.UserPrincipal;
-import rush.rush.service.article.comment.DeleteCommentService;
+import rush.rush.service.comment.DeleteCommentService;
 
 @RestController
 @RequestMapping("/comments")
